+++ include/tvm/tirx/op_attr_types.h
@@ -80,7 +80,7 @@ enum class ScriptDtypePrintLocation : int {
   kLast = 2,
 };
 
-using TScriptDtypePrintLocation = Integer;
+using TScriptDtypePrintLocation = int64_t;
 
 /*!
  * \brief The effect type of the call.
@@ -149,7 +149,7 @@ inline std::ostream& operator<<(std::ostream& os, CallEffectKind side_effect) {
 }
 
 /*! \brief Use integer to record the kind. */
-using TCallEffectKind = Integer;
+using TCallEffectKind = int64_t;
 
 }  // namespace tirx
 }  // namespace tvm
